+++ assembly/ByteArray/Hex.ts
@@ -12,10 +12,10 @@ export const mapping: string[] = [];
 const alphabet: string = '0123456789abcdef';
 
 // 生成映射表
-for (let i: u8 = 0; i < 16; ++i) {
-  const offset: u8 = i * 16;
+for (let i: i32 = 0; i < 16; ++i) {
+  const offset: i32 = i * 16;
 
-  for (let j: u8 = 0; j < 16; ++j) {
+  for (let j: i32 = 0; j < 16; ++j) {
     mapping[offset + j] = alphabet.charAt(i) + alphabet.charAt(j);
   }
 }
